+++ src/main/java/com/checkcheck/ecoreading/config/NaverConfig.java
@@ -1,0 +1,24 @@
+//package com.checkcheck.ecoreading.config;
+//
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.boot.context.properties.ConfigurationProperties;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.context.annotation.Configuration;
+//
+//@Configuration
+//public class NaverConfig {
+//    @Value("${naver-property.clientId}")
+//    private String clientId;
+//    @Value("{naver-property.clientSecret}")
+//    private String clientSecret;
+//
+//    @Bean
+//    public String naverClientId() {
+//        return clientId;
+//    }
+//
+//    @Bean
+//    public String naverClientSecret() {
+//        return clientSecret;
+//    }
+//}
