+++ src/main/java/com/checkcheck/ecoreading/domain/boards/controller/BookController.java
@@ -1,69 +1,44 @@
-package com.checkcheck.ecoreading.domain.boards.controller;
-
-import com.checkcheck.ecoreading.domain.boards.dto.BookDTO;
-import com.checkcheck.ecoreading.domain.boards.dto.NaverResultDTO;
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.springframework.http.RequestEntity;
-import org.springframework.http.ResponseEntity;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.client.RestTemplate;
-import org.springframework.web.util.UriComponentsBuilder;
-
-import java.net.URI;
-import java.util.List;
-
-@Controller
-public class BookController {
-    @GetMapping("/board/bookSearch")
-    public String search(@RequestParam String text, Model model) {
-        //네이버 검색 API 요청
-        String clientId = "7mvqmOEqDyA7VG3qoDEa";
-        String clientSecret = "LHramuXw91";
-
-        // JSON 결과
-        // String apiURL = "https://openapi.naver.com/v1/search/book.json?query="+ text;
-        URI uri = UriComponentsBuilder
-                .fromUriString("https://openapi.naver.com")
-                .path("/v1/search/book.json")
-                .queryParam("query", text)
-                .queryParam("display", 10)
-                .queryParam("start", 1)
-                .queryParam("sort", "sim")
-                .encode()
-                .build()
-                .toUri();
-
-        // Spring 요청 제공 클래스  (Request)
-        RequestEntity<Void> request = RequestEntity.get(uri)
-                .header("X-Naver-Client-Id", clientId)
-                .header("X-Naver-Client-Secret", clientSecret)
-                .build();
-
-        // Spring 제공 restTemplate (Response)
-        RestTemplate restTemplate = new RestTemplate();
-        ResponseEntity<String> response = restTemplate.exchange(request, String.class);
-
-        // JSON 파싱 (JSON 문자열을 객체로 만듦, 문서화하기)
-        ObjectMapper om = new ObjectMapper();
-        NaverResultDTO resultDTO = null;
-
-        try {
-            resultDTO = om.readValue(response.getBody(), NaverResultDTO.class);
-        } catch (JsonMappingException e) {
-            e.printStackTrace();
-        } catch (JsonProcessingException e) {
-            e.printStackTrace();
-        }
-
-        List<BookDTO> books = resultDTO.getItems(); // books를
-        System.out.println(books);
-        model.addAttribute("books", books);
-
-        return "/content/user/bookSearch";
-    }
-}
+//package com.checkcheck.ecoreading.domain.boards.controller;
+//
+//import com.checkcheck.ecoreading.domain.boards.dto.BookDTO;
+//import com.checkcheck.ecoreading.domain.boards.dto.NaverResultDTO;
+//import com.checkcheck.ecoreading.domain.boards.service.BookService;
+//import com.fasterxml.jackson.core.JsonProcessingException;
+//import com.fasterxml.jackson.databind.JsonMappingException;
+//import com.fasterxml.jackson.databind.ObjectMapper;
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.beans.factory.annotation.Value;
+//import org.springframework.http.RequestEntity;
+//import org.springframework.http.ResponseEntity;
+//import org.springframework.stereotype.Controller;
+//import org.springframework.ui.Model;
+//import org.springframework.web.bind.annotation.GetMapping;
+//import org.springframework.web.bind.annotation.PostMapping;
+//import org.springframework.web.bind.annotation.RequestBody;
+//import org.springframework.web.bind.annotation.RequestParam;
+//import org.springframework.web.client.RestTemplate;
+//import org.springframework.web.util.UriComponentsBuilder;
+//
+//import java.net.URI;
+//import java.util.List;
+//
+//@Controller
+//@RequiredArgsConstructor
+//public class BookController {
+//    private final BookService bookService;
+//
+//    @GetMapping("/board/bookSearch")
+//    public String search(@RequestParam String text, Model model) {
+//        List<BookDTO> books = bookService.searchBooks(text);
+//        System.out.println(books);
+//        model.addAttribute("books", books);
+//
+//        return "/content/user/bookSearch";
+//    }
+//
+//    @PostMapping("/board/new")
+//    public String fillBook(@RequestBody BookDTO bookDTO){
+//        System.out.println(bookDTO);
+//        return "/content/user/boardAddForm";
+//    }
+//}
