+++ src/main/resources/templates/content/user/bookSearch.html
@@ -12,35 +12,36 @@
         <th>이미지</th>
         <th>ISBN</th>
         <th>제목</th>
-        <th>저자/출판사</th>
+        <th>저자</th>
+        <th>출판사</th>
         <th>발행일</th>
         <th>선택</th>
     </tr>
     <tr th:each="book : ${books}">
-        <td><img th:src="${book.image}" th:alt="${book.title}" width="100"></td>
-        <td th:text="${book.isbn}"></td>
-        <td><a th:href="${book.link}" th:text="${book.title}"></a></td>
-        <td th:text="${book.author} + '/' + ${book.publisher}"></td>
-        <td th:text="${book.pubdate}"></td>
-        <td><input type="checkbox" id="myCheckbox" onclick="processCheckbox()" th:value="${book.isbn}"></td>
+        <td><img th:src="${book.image}" id="image" th:alt="${book.image}" width="100"></td>
+        <td th:text="${book.isbn}" th:id="isbn" ></td>
+        <td><a th:text="${book.title}" th:name="title"></a></td>
+        <td th:text="${book.author}" th:name="author"></td>
+        <td th:text="${book.publisher}" th:name="publisher"></td>
+        <td th:text="${book.pubdate}" th:name="pubdate"></td>
+        <td>
+            <form th:action="@{/board/bookSearch}" th:object="${BookDTO}" th:method="post">
+            <button type="submit" class="btn btn-dark mx-1">선택</button>
+            <input type="hidden" name="title" th:value="${book.title}" />
+            <input type="hidden" name="isbn" th:value="${book.isbn}" />
+            <input type="hidden" name="image" th:value="${book.image}" />
+            <input type="hidden" name="author" th:value="${book.author}" />
+            <input type="hidden" name="publisher" th:value="${book.publisher}" />
+            <input type="hidden" name="pubdate" th:value="${book.pubdate}" />
+            <input type="hidden" name="description" th:value="${book.description}" />
+            </form>
+        </td>
     </tr>
 </table>
 </body>
 
 <script>
-    function processCheckbox(checkbox) {
-        // 체크박스의 상태를 확인
-        checkbox = document.getElementById('myCheckbox');
-        var isChecked = checkbox.checked;
-        var isbn = checkbox.value;
 
-        if (checkbox.checked) {
-            // 선택되었을 때 실행할 코드
-            alert('책이 선택되었습니다. ISBN: ' + isbn);
-            // 여기에서 창을 닫는 등의 작업을 수행할 수 있습니다.
-            window.close();
-        }
-    }
 </script>
 
 </html>
